Check immediate field size before determining overflow.
Before this patch, the following (legal!) code will cause an overflow
error during assembly.

    #define HI(z) (((z) >> 16) & $FFFF)
    #define LO(z) ((z) & $FFFF)

        *=$2233

    SYM:
        rts

        lda #HI(SYM)
        pha
        lda #LO(SYM)	; generates Overflow error
        pha

This happens because overflow is checked for ahead of the assembler
checking to see if a two-byte immediate field is allowed for the
instruction.  This patch re-arranges the control flow logic so that the
immediate field width capabilities are checked for prior to checking for
overflow.

I've also confirmed that this works even if we set *=$112233, thus
exercising the case for a 24-bit label value.

This patch DOES NOT relate to relocation generation, however.  Thus, if
you're generating an o65 file, do not expect the above code to produce
relocations for SYM.